test(validation): extract CDI classifier cases to focused module
